Updated NuGet packages, added Directory.Build.props, consolidated all projects to NET6 and NET4.8 for forms, fixed build errors.

+++ Win7Samples/winui/obscure/Program.cs
@@ -74,7 +74,7 @@ void Obscure_EnsureTimerStopped(HWND hwnd)
 {
 	if (g_fTimerActive)
 	{
-		KillTimer(hwnd, (IntPtr)1);
+		KillTimer(hwnd, 1);
 		g_fTimerActive = false;
 		InvalidateRect(hwnd, default, false);
 		SetWindowText(hwnd, "Covered (Paused)");
@@ -88,7 +88,7 @@ void Obscure_EnsureTimerStopped(HWND hwnd)
 * Otherwise, invalidate our rectangle so we will redraw with the new time.
 *
 *****************************************************************************/
-void Obscure_TimerProc(HWND hwnd, uint uiMsg, IntPtr idTimer, uint tm)
+void Obscure_TimerProc(HWND hwnd, uint uiMsg, nuint idTimer, uint tm)
 {
 	/*
 	* If the client area is totally obscured, then stop the timer so we don't waste any more CPU.
@@ -122,7 +122,7 @@ void Obscure_EnsureTimerRunning(HWND hwnd)
 {
 	if (!g_fTimerActive)
 	{
-		SetTimer(hwnd, (IntPtr)1, 100, Obscure_TimerProc);
+		SetTimer(hwnd, 1, 100, Obscure_TimerProc);
 		g_fTimerActive = true;
 	}
 }
